requirements.txt
boto3==1.14.48
botocore==1.17.48
click==7.1.2
docutils==0.15.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
numpy==1.19.1
opencv-python==4.4.0.42
Pillow==7.2.0
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
scipy
torchvision==0.8.0